+++ rollup.config.ts
@@ -17,7 +17,7 @@ import type {
   OutputOptions,
 } from 'rollup';
 
-type BuildMode = 'inline' | 'slim' | 'node';
+type BuildMode = 'inline' | 'slim' | 'bundle';
 
 type BuildName = string;
 
@@ -44,6 +44,7 @@ function pluginCopyWasm(): Plugin {
     name: 'copy-wasm',
     resolveImportMeta: () => `""`,
     generateBundle() {
+      fs.mkdirSync(DIST_DIR, { recursive: true });
       fs.copyFileSync(
         path.resolve('./node_modules/p2panda-js/lib/p2panda.wasm'),
         path.resolve(`${DIST_DIR}/p2panda.wasm`),
@@ -55,8 +56,8 @@ function pluginCopyWasm(): Plugin {
 // Returns the name of the sub-directory which will be created in the target
 // folder for each build.
 function getBuildName({ format, mode }: Config): BuildName {
-  if (mode === 'node') {
-    return 'node';
+  if (mode === 'bundle') {
+    return `${format}-bundle`;
   } else if (mode === 'inline') {
     return format;
   } else {
@@ -81,25 +82,25 @@ function getOutputs({ format, mode }: Config): OutputOptions[] {
     sourcemap,
   });
 
-  // Provide a minified version for non-NodeJS builds
-  if (mode !== 'node') {
-    result.push({
-      name: pkg.name,
-      file: `${DIST_DIR}/${dirName}/${BUILD_FILE_NAME}.min.js`,
-      format,
-      sourcemap,
-      plugins: [pluginTerser()],
-    });
-  }
+  // Provide a minified version for all builds
+  result.push({
+    name: pkg.name,
+    file: `${DIST_DIR}/${dirName}/${BUILD_FILE_NAME}.min.js`,
+    format,
+    sourcemap,
+    // @ts-expect-error: TypeScript misinterprets the module configuration
+    plugins: [pluginTerser()],
+  });
 
   return result;
 }
 
-function getPlugins({ format, mode }: Config): Plugin[] {
+function getPlugins({ mode }: Config): Plugin[] {
   const result: Plugin[] = [];
 
   // Convert external CommonJS- to ES6 modules
   result.push(
+    // @ts-expect-error: TypeScript misinterprets the module configuration
     pluginCommonJS({
       extensions: ['.js', '.ts'],
     }),
@@ -107,7 +108,7 @@ function getPlugins({ format, mode }: Config): Plugin[] {
 
   // In umd builds we're bundling the dependencies as well, we need this plugin
   // here to help locating external dependencies
-  if (format === 'umd') {
+  if (mode === 'bundle') {
     result.push(
       pluginNodeResolve({
         // Use the "browser" module resolutions in the dependencies' package.json
@@ -116,27 +117,35 @@ function getPlugins({ format, mode }: Config): Plugin[] {
     );
   }
 
+  // graphql-web-lite provides an alias package that can be swapped in for
+  // the standard graphql package in client-side applications. It aims to
+  // reduce the size of imports that are in common use by GraphQL clients
+  // and users, while still providing most graphql exports that are used
+  // in other contexts.
+  const entries = [{ find: 'graphql', replacement: 'graphql-web-lite' }];
+
   // Whenever we want to build a "slim" version of shirokuma we have to import
   // the "slim" version of p2panda-js.
   //
   // The "slim" version does not contain the WebAssembly inlined (as a base64
   // string) and is therefore smaller.
   if (mode === 'slim') {
-    result.push(
-      pluginAlias({
-        entries: [{ find: 'p2panda-js', replacement: 'p2panda-js/slim' }],
-      }),
-    );
+    entries.push({ find: 'p2panda-js', replacement: 'p2panda-js/slim' });
   }
+
+  result.push(
+    // @ts-expect-error: TypeScript misinterprets the module configuration
+    pluginAlias({
+      entries,
+    }),
+  );
 
   // Compile TypeScript source code to JavaScript
+  // @ts-expect-error: TypeScript misinterprets the module configuration
   result.push(pluginTypeScript());
 
-  // We only need to copy this once, let's pick NodeJS build for that (since it
-  // only gets build once)
-  if (mode === 'node') {
-    result.push(pluginCopyWasm());
-  }
+  // We only need to copy this once
+  result.push(pluginCopyWasm());
 
   return result;
 }
@@ -153,13 +162,12 @@ function config({ format, mode }: Config): RollupOptions[] {
   // Determine plugins which will be used to process this build
   const plugins = getPlugins({ format, mode });
 
-  // Treat all npm dependencies as external, we don't want to include them in
-  // our bundle.
+  // We build modules in two versions: 1. with no 3rd party dependencies included
+  // and 2. bundled inside of them.
   //
-  // Only "umd" bundles contain all dependencies bundled as well. That's a
-  // shame, but we hope that this will account for the "quick" uses of
-  // shirokuma.
-  const external = format === 'umd' ? [] : Object.keys(pkg.dependencies);
+  // Bundling them is somewhat a shame, but we hope that this will account for
+  // the "quick" uses of shirokuma.
+  const external = mode === 'bundle' ? [] : Object.keys(pkg.dependencies);
 
   // Package build
   result.push({
@@ -185,16 +193,16 @@ function config({ format, mode }: Config): RollupOptions[] {
 
 export default [
   ...config({
-    format: 'umd',
-    mode: 'inline',
-  }),
-  ...config({
     format: 'cjs',
     mode: 'inline',
   }),
   ...config({
     format: 'cjs',
     mode: 'slim',
+  }),
+  ...config({
+    format: 'cjs',
+    mode: 'bundle',
   }),
   ...config({
     format: 'esm',
@@ -205,7 +213,7 @@ export default [
     mode: 'slim',
   }),
   ...config({
-    format: 'cjs',
-    mode: 'node',
+    format: 'esm',
+    mode: 'bundle',
   }),
 ];